+++ src/components/icons/column.tsx
@@ -1,0 +1,16 @@
+const ColumnIcon = () => {
+  return (
+    <svg
+      className="icon"
+      viewBox="0 0 1024 1024"
+      version="1.1"
+      xmlns="http://www.w3.org/2000/svg"
+      width="20"
+      height="20"
+    >
+      <path d="M96 128v704a64 64 0 0 0 64 64h704a64 64 0 0 0 64-64V128H96z m64 128h320v576H160V256z m704 576H544V256h320v576z"></path>
+    </svg>
+  )
+}
+
+export default ColumnIcon