+++ src/components/multiFile/index.tsx
@@ -1,17 +1,26 @@
 import React, { useMemo, useState } from 'react'
-import { Checkbox, Drawer, Popconfirm, Popover, Tree, Upload } from 'antd'
-import FileTitleModal from '../fileTitleModal'
+import { Checkbox, Drawer, Tree, Upload } from 'antd'
+import FileTitleModal from './fileTitleModal'
 import MultiFileIcon from '../icons/multiFile'
 import { arrToTree, FileData } from '@/utils/multiFile'
 import styles from './index.module.less'
-import MoreIcon from '../icons/more'
-import DeleteFileIcon from '../icons/deleteFile'
-import EditFileIcon from '../icons/editFile'
 import { useFile } from '@/context/file'
 import { ACTION_TYPE } from '@/context/file/reducer'
 import UploadIcon from '../icons/upload'
 import DownloadIcon from '../icons/download'
 import iconText from '@/assets/imgs/iconText.png'
+import TitleRender from './titleRender'
+
+interface DragData extends FileData {
+  pos: string
+}
+
+export interface DropInfo {
+  node: DragData
+  dragNode: DragData
+  dropPosition: number
+  dropToGap: boolean
+}
 
 const MultiFile: React.FC = () => {
   const { file, dispatch } = useFile()
@@ -19,17 +28,15 @@ const MultiFile: React.FC = () => {
   const [isEditor, setIsEditor] = useState(false) // edit status
   const [isTitleModalOpen, setIsTitleModalOpen] = useState(false) // title modal open status
   const [editTitleData, setEditTitleData] = useState<FileData>({} as FileData) // Current edit title data
+
   const treeData = useMemo(
     () => arrToTree(file.multiFileData),
     [file.multiFileData]
   )
-  // Number of files(not folders)
-  const fileCount = useMemo(
-    () =>
-      file.multiFileData.reduce((acc, data) => acc + (data.isLeaf ? 1 : 0), 0),
-    [file.multiFileData]
-  )
 
+  /**
+   * click file
+   */
   const onSelect = (_selectedKeys: React.Key[], info: { node: FileData }) => {
     if (info.node.isLeaf) {
       dispatch({
@@ -39,19 +46,14 @@ const MultiFile: React.FC = () => {
     }
   }
 
-  const deleteFile = (key: string) => {
+  /**
+   * drag and drop file
+   */
+  const onDrop = (info: DropInfo) => {
     dispatch({
-      type: ACTION_TYPE.DELTE_FILE,
-      payload: key
-    })
-  }
-
-  const addFile = (parent: string, isLeaf: boolean) => {
-    dispatch({
-      type: ACTION_TYPE.ADD_FILE,
+      type: ACTION_TYPE.SORT_FILE,
       payload: {
-        isLeaf,
-        parent
+        info
       }
     })
   }
@@ -65,80 +67,6 @@ const MultiFile: React.FC = () => {
       }
     })
     setIsTitleModalOpen(false)
-  }
-
-  const titleRender = (fileData: FileData) => {
-    return (
-      <div
-        onClick={(e) => {
-          e.stopPropagation()
-        }}
-        className={styles.titleContainer}
-      >
-        <span className={styles.titleText}>{fileData.title}</span>
-        <span
-          onClick={() => {
-            setEditTitleData(fileData)
-            setIsTitleModalOpen(true)
-          }}
-          className={styles.mutilFileIcon}
-        >
-          <EditFileIcon />
-        </span>
-        {fileCount > 1 && fileData.key !== file.curKey && (
-          <Popconfirm
-            title="Are you sure to delete this file?"
-            onConfirm={() => deleteFile(fileData.key)}
-            okText="Yes"
-            cancelText="No"
-          >
-            <span className={styles.mutilFileIcon}>
-              <DeleteFileIcon />
-            </span>
-          </Popconfirm>
-        )}
-
-        <Popover
-          content={
-            <div className={styles.titleOperate}>
-              <div
-                className={styles.titleLine}
-                onClick={() => addFile(fileData.parent, true)}
-              >
-                Add file at the current level
-              </div>
-              <div
-                className={styles.titleLine}
-                onClick={() => addFile(fileData.parent, false)}
-              >
-                Add folder at current level
-              </div>
-              {!fileData.isLeaf && (
-                <>
-                  <div
-                    className={styles.titleLine}
-                    onClick={() => addFile(fileData.key, true)}
-                  >
-                    Add file at the sub level
-                  </div>
-                  <div
-                    className={styles.titleLine}
-                    onClick={() => addFile(fileData.key, false)}
-                  >
-                    Add folder at sub level
-                  </div>
-                </>
-              )}
-            </div>
-          }
-          trigger="click"
-        >
-          <span className={styles.mutilFileIcon}>
-            <MoreIcon />
-          </span>
-        </Popover>
-      </div>
-    )
   }
 
   const downloadFile = () => {
@@ -215,10 +143,20 @@ const MultiFile: React.FC = () => {
               if (!isEditor) {
                 return <div className={styles.titleText}>{node.title}</div>
               }
-              return titleRender(node)
+              return (
+                <TitleRender
+                  fileData={node}
+                  setEditTitleData={(data) => {
+                    setEditTitleData(data)
+                    setIsTitleModalOpen(true)
+                  }}
+                />
+              )
             }}
             selectedKeys={[file.curKey]}
             showLine
+            draggable={isEditor}
+            onDrop={onDrop}
             rootStyle={{
               overflowY: 'auto',
               flex: '1'