+++ src/context/file/index.tsx
@@ -8,6 +8,8 @@ import {
 } from 'react'
 import { FileAction, FileReducer, fileReducer, FileState } from './reducer'
 import { getFileStorageData } from '@/utils/storage'
+import { MD, toc } from '@/utils/md'
+import { VIEW_STATE } from '@/utils/constants'
 
 const FileContext = createContext<{
   file: FileState
@@ -16,14 +18,18 @@ const FileContext = createContext<{
 FileContext.displayName = 'FileContext'
 
 const fileInitState = getFileStorageData()
+const initHtmlStr = MD.render(fileInitState.currentData.content)
+
 export const FileProvider: FC<{ children: ReactNode }> = ({ children }) => {
   const [file, dispatch] = useReducer<FileReducer>(fileReducer, {
     content: fileInitState.currentData.content,
+    htmlStr: initHtmlStr,
+    viewState: VIEW_STATE.ALL,
+    titleList: toc.get(),
     state: fileInitState.currentData.state,
     curKey: fileInitState.currentKey,
     multiFileData: fileInitState.storageData.fileData
   })
-
   return (
     <FileContext.Provider value={{ file, dispatch }}>
       {children}