+++ src/views/editor/index.tsx
@@ -1,7 +1,5 @@
-import React, { useMemo, useRef, useState } from 'react'
-import { MD, toc } from '@/utils/md'
+import React, { useState } from 'react'
 import { clearScrollMap, editorScroll, previewScroll } from '@/utils/scroll'
-import { ITitle } from '@/utils/toc'
 import { useResizeEvent } from '@/hooks/event'
 
 import { Layout } from 'antd'
@@ -10,26 +8,17 @@ import styles from './index.module.less'
 import { useFile } from '@/context/file'
 import { ACTION_TYPE } from '@/context/file/reducer'
 import editorbg from '@/assets/imgs/editorbg.png'
+import { VIEW_STATE } from '@/utils/constants'
 const { Header, Content } = Layout
 
+let contentTimer: NodeJS.Timeout
 const Editor: React.FC = () => {
   const { file, dispatch } = useFile()
-  const editorRef = useRef<HTMLTextAreaElement>(null) // edit area ref
-  const previewRef = useRef<HTMLDivElement>(null) // preview area ref
+  const [isInput, setIsInput] = useState(false)
+  const [editorEl, setEditorEl] = useState<HTMLTextAreaElement | null>(null) // edit area element
+  const [previewEl, setPreviewEl] = useState<HTMLDivElement | null>(null) // preview area element
   const [showToc, setShowToc] = useState(false) // toc display
-  const [titleList, setTitleList] = useState<ITitle[]>([]) // title list
-
   useResizeEvent(clearScrollMap)
-
-  /**
-   * Get the converted html string and get the title list
-   */
-  const htmlStr = useMemo(() => {
-    toc.clear()
-    const str = MD.render(file.content)
-    setTitleList(toc.get())
-    return str
-  }, [file.content])
 
   /**
    * change markdown content
@@ -41,6 +30,17 @@ const Editor: React.FC = () => {
       payload: content
     })
     clearScrollMap()
+    setIsInput(true)
+    if (editorEl && previewEl) {
+      const { scrollHeight, scrollTop, clientHeight } = editorEl
+      if (scrollHeight - clientHeight < scrollTop + 6) {
+        previewEl.scrollTop = previewEl.scrollHeight
+      }
+    }
+    contentTimer && clearTimeout(contentTimer)
+    contentTimer = setTimeout(() => {
+      setIsInput(false)
+    }, 50)
   }
 
   /**
@@ -49,11 +49,13 @@ const Editor: React.FC = () => {
    */
   const scrollHandle = (el: HTMLInputElement) => {
     const nodeName = el.nodeName
-    if (editorRef.current && previewRef.current) {
+    if (editorEl && previewEl) {
       if (nodeName === 'TEXTAREA') {
-        editorScroll(editorRef.current, previewRef.current)
-      } else {
-        previewScroll(editorRef.current, previewRef.current)
+        editorScroll(editorEl, previewEl)
+      } else if (nodeName === 'DIV') {
+        if (!isInput) {
+          previewScroll(editorEl, previewEl)
+        }
       }
     }
   }
@@ -61,35 +63,39 @@ const Editor: React.FC = () => {
   return (
     <Layout className={styles.layout}>
       <Header className={styles.layoutHeader}>
-        <Toolbar
-          editor={editorRef.current}
-          setShowToc={() => setShowToc((v) => !v)}
-        />
+        <Toolbar editor={editorEl} setShowToc={() => setShowToc((v) => !v)} />
       </Header>
       <Content>
         <div className={styles.container}>
-          <textarea
-            className={`${styles.content} ${styles.editorTextarea}`}
-            ref={editorRef}
-            value={file.content}
-            onInput={(e) => {
-              changeMdContent((e.target as HTMLInputElement).value)
-            }}
-            style={{ backgroundImage: `url(${editorbg})` }}
-            onScroll={(e) => scrollHandle(e.target as HTMLInputElement)}
-          ></textarea>
-          <div
-            id="write"
-            ref={previewRef}
-            className={`${styles.content} ${styles.write}`}
-            dangerouslySetInnerHTML={{ __html: htmlStr }}
-            onScroll={(e) => scrollHandle(e.target as HTMLInputElement)}
-          ></div>
+          {(file.viewState === VIEW_STATE.ALL ||
+            file.viewState === VIEW_STATE.EDITOR) && (
+            <textarea
+              className={`${styles.content} ${styles.editorTextarea}`}
+              ref={setEditorEl}
+              value={file.content}
+              onInput={(e) => {
+                changeMdContent((e.target as HTMLInputElement).value)
+              }}
+              style={{ backgroundImage: `url(${editorbg})` }}
+              onScroll={(e) => scrollHandle(e.target as HTMLInputElement)}
+            ></textarea>
+          )}
+          {(file.viewState === VIEW_STATE.ALL ||
+            file.viewState === VIEW_STATE.PREVIEW) && (
+            <div
+              id="write"
+              ref={setPreviewEl}
+              className={`${styles.content} ${styles.write}`}
+              dangerouslySetInnerHTML={{ __html: file.htmlStr }}
+              onScroll={(e) => scrollHandle(e.target as HTMLInputElement)}
+            ></div>
+          )}
+
           {showToc && (
             <div className={styles.toc}>
               <div className={styles.tocTitle}>TOC</div>
               <div>
-                {titleList.map(({ val, level }, index) => {
+                {file.titleList.map(({ val, level }, index) => {
                   const fontSize = ((7 - level) / 10) * 40
 
                   return (